Added CNV and Fusion report. Added collated report output.
We'll have some panels that include CNV and / or Fusion data with the
SNV / Indel data that was originally considered.  Added ability to add
those data to the report if desired.  Also, added a singular, combined
report that accomodates all three variant types.  Pipeline working and
ready for testing / tweaking.