Modified to handle new refGene file.
Annovar will tack on the name of the refGene file to the elements that
are annotated with it (gene annotation).  So, gene name field becomes
'Gene.refGeneWithVer'.  If we change things, then most of this script
breaks.  Chnaged the way the header is handled so that we don't have
that problem.