Updated parsed data struct from array to dict.
When we change any annotation components (especially adding / removing)
there is a chance that the columns move around and as such things break.
Converted the parsed data to an array of dicts so that we can always
just reference the data by column name and don't have to worry about
things moving around.