Fixed VAF and coverage output. Added pop freq filter.
Added the population frequency option to this main pipeline script so
that we can adjust here instead of in `moma.pl`. Also fixed the VAF calc
and the ref / alt reads parsing. Was not always working depending on the
VCF, so added a more robust dict method of parsing this.