Added protocol num and id validation.
Since the Protocol Number, Treatment Patient ID, and Specimen ID strings
are important and may be typed by hand, added some simple regex
validation to make sure that we will get the correct strings.  As more
of these are processed, will probably have to refine, but should be good
to start with now.