Added moma2rave.py call, make arg sections, and refactored some logging.
For Biobank and Biobank Pilot studies, need a custom CSV file from a
MOMA report to be uploaded. Added that option to this pipeline script to
call the `moma2rave.py` script to generate the report and store it in
the output dir.

Since we are starting to build up a long list of args, added some arg
gruops that we can use for organization and better output with running
with the `--help` option.

Refactored some of the header and log strings to be a bit more organized
and render better. Also reorganized the initial structure for running
the `--noanno` option (was missing the proper `outdir_path` var and
getting an error.)