Re-vamped get_sbs_metrics and simply_vcf
Needed to deal with not having a full VCF Extractor program for the
mutational signature script.  For the rest, we're using
`simplify_vcf.pl`, and so I'd like to use that to do this as well. I
added a CSV output method, which will be what we need to get the data
into the SBS script.  At the same time, in order to keep this SBS script
version different than the full version, I have renamed it.