Fixed logging and added sample name to logfile.
Log file creation is happening too early, so when we are running the
help ,for example, a logfile stump is created and then left behind.
Also, when running on an HPC in a swarm fashion, the logfile is getting
misplaced and not making it with the data.  Plus the quiet function is
not quite workig correctly.  All this is fixed now, as we're making the
logfile just beofre we actually need it now.