Updated gene list and added new info.
Needed to get strand information so that we can indicate which strand
the gene is on as part of the MAF standard.  Also, as part of this new
mapping, there are about 40 genes, which have had their names changed
(still same Entrez ID, RefSeq ID, and whatnot). Since I do need to
account for both cases, I'm going to keep the old gene entries in here,
and just add the new ones.  Later, once we not longer need to map the
older gene names, I can remove the old entries from this file.