Fixed issue with some WES VCFs and VAF determination
The Ts/Tv and Deamination score scripts require VAF extraction from the
VCF. But, some of the WES files I get seem to be a bit wonky with that
data and the results are null or 0. Fixed the way this is handled so
that downstream processing isn't affected by these entries.