Working on MuTect2 tumor:normal parsing.
The WES data we're getting can be tumor only, run through the Haplotype
Caller pipeline, *or* from a tumor / normal pairing, run through
MuTect2. I was not prepared for the latter, and have written a parser to
strip out the normal data, and updated the pipeline to accomodate it
all.

Also updated the gitignore and added some new test VCFs.