Added ability to read gzipped / bgzipped files.
Some of the larger data may come in automatically as bgzipped or just
gzipped VCF files. Added some functionlity to handle this by
decompressing the data, reading and processing it, and then dumping the
decompressed VCF to save space.  I think that in all cases the data size
won't be so large that using some temp file space for this is a problem.